Made a bunch of internal calls inline

+++ AnalyzerFFT.h
@@ -187,7 +187,7 @@ void AnalyzerFFTSingle(uint16_t *ana, const float *samp, int inc, int bits)
 
     for (auto i = 0; i < full; ++i)
     {
-        fft_x86[i][0] = (int32_t)(fmaxf(fminf(*samp, 1.0f), -1.0f) * SHRT_MAX) << 12;
+        fft_x86[i][0] = (int32_t)(fmaxf(fminf(*samp, 1.0f), -1.0f) * (float)SHRT_MAX) << 12;
         samp += inc;
         fft_x86[i][1] = 0;
     }
